sem VI/AI-Carrer-Project/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
python-docx==1.1.2
google-generativeai
python-dotenv==1.0.1
pydantic==2.9.2